scraping/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
supabase==2.8.0
psycopg2-binary==2.9.9